Added BaseViewModel in App.xaml and added the viewmodel as a sharedviewmodel for every sub study creation pages
