+++ StudyConfigurationUI/StudyConfigurationUI/View/Pages/StudyCreationPages/StudyGeneralSettingsPage.xaml.cs
@@ -6,8 +6,10 @@
 
 using System;
 using Windows.Storage.Pickers;
+using Windows.UI.Popups;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
+using StudyConfigurationUI.ViewModel;
 
 #endregion
 
@@ -20,9 +22,13 @@ namespace StudyConfigurationUI.View.Pages.StudyCreationPages
     /// </summary>
     public sealed partial class StudyGeneralSettingsPage : Page
     {
+
+        private StudyCreationPageViewModel _viewModel;
         public StudyGeneralSettingsPage()
         {
             InitializeComponent();
+            _viewModel = App.StudyViewModel;
+            DataContext = _viewModel;
         }
 
         private async void SelectFile_OnClick(object sender, RoutedEventArgs e)
@@ -43,9 +49,35 @@ namespace StudyConfigurationUI.View.Pages.StudyCreationPages
             }
         }
 
-        private void SubmitFile_OnClick(object sender, RoutedEventArgs e)
+        private async void SubmitBut_OnClick(object sender, RoutedEventArgs e)
         {
-            throw new NotImplementedException();
+            if (!_viewModel.ValidateStudy())
+            {
+                var errorMsg =
+                    new MessageDialog("Study has not been configured correctly. Please validate your settings again")
+                    {
+                        Title = "Error"
+                    };
+                return;
+            }
+
+
+            var dialog = new MessageDialog(
+                "Are you sure you want to submit the Study?")
+            { Title = "Submit Phase." };
+            dialog.Commands.Add(new UICommand("Yes") { Id = 0 });
+            dialog.Commands.Add(new UICommand("No") { Id = 1 });
+
+            dialog.DefaultCommandIndex = 1;
+            dialog.CancelCommandIndex = 1;
+
+            var choice = await dialog.ShowAsync();
+
+            if (choice.Id.Equals(0))
+            {
+                //Todo Submit method
+            }
         }
+
     }
 }
